+++ src/components/App.js
@@ -1,12 +1,13 @@
 import Nav from "./Nav";
 
+
 function App({ children }) {
-  return (
-    <div>
-      <Nav></Nav>
-      <div>{children}</div>
-    </div>
-  );
+	return (
+		<div>
+			<Nav></Nav>
+			<div>{children}</div>
+		</div>
+	);
 }
 
 export default App;
