--- src/components/BestList.module.css
@@ -1,31 +1,0 @@
-@font-face {
-  font-family: "Pretendard Variable";
-  src: url("../../public/Pretendard-Regular.ttf");
-}
-
-.label {
-  font-family: "Pretendard Variable";
-  font-size: 20px;
-  font-weight: 700;
-  line-height: 32px;
-  text-align: left;
-  color: #111827;
-}
-
-.bestList {
-  display: flex;
-  justify-content: center;
-  gap: 24px;
-}
-
-.bestListBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 16px;
-  width: fit-content;
-}
-
-.wrapper {
-  display: flex;
-  justify-content: center;
-}
