+++ src/components/addItem/ItemInfo.js
@@ -1,0 +1,10 @@
+const ItemInfo = () => {
+  return(
+    <>
+      <h1>상품 소개</h1>
+      <textarea placeholder="상품 소개를 입력해주세요"></textarea>
+    </>
+  )
+};
+
+export default ItemInfo;
