+++ src/components/addItem/ItemName.js
@@ -1,0 +1,10 @@
+const ItemName = () => {
+  return(
+    <>
+      <h1>상품명</h1>
+      <textarea placeholder="상품명을 입력해주세요"></textarea>
+    </>
+  )
+}
+
+export default ItemName;
