+++ src/components/addItem/UploadImage.js
@@ -1,0 +1,11 @@
+const UploadImage = () => {
+  return(
+    <>
+      <h2>상품 이미지</h2>
+      <h2>이미지 박스(임의)</h2>
+    </>
+  )
+}
+
+
+export default UploadImage;
