+++ src/components/itemList/Card.module.css
@@ -1,6 +1,6 @@
 @font-face {
   font-family: "Pretendard Variable";
-  src: url("../../public/Pretendard-Regular.ttf");
+  src: url("../../../public/Pretendard-Regular.ttf");
 }
 
 .cardImg {
