+++ src/main.js
@@ -1,6 +1,7 @@
 import { BrowserRouter, Routes, Route } from "react-router-dom";
 import App from "./components/App";
 import ItemPage from "./pages/ItemPage";
+import AddItemPage from "./pages/AddItemPage";
 
 function Main() {
   return (
@@ -8,6 +9,7 @@ function Main() {
       <App>
         <Routes>
           <Route path="/items" element={<ItemPage />} />
+          <Route path="/additem" element={<AddItemPage />}/>
         </Routes>
       </App>
     </BrowserRouter>
